Fix an issue with subscriptions silently breaking
If the app supports background operation, the multicast sockets are
shutdown as the app moves to background. The result of this is that when
the app moves to the foreground, subscriptions are invalid.

To fix this, subscription models should have a resubscribe method which
resubscribes without creating a new subscription object.